Move the csv decoder instance out of Lorre main body
